finished some css styling for the book cards
