+++ frontend/src/components/BookCards/BookCard.css
@@ -1,0 +1,45 @@
+.books-list {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    overflow-y: scroll;
+}
+
+.book-card {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 160px;
+    height: 400px;
+    margin: 10px;
+    border: 1px solid #f7a4f0;
+    background-color: #f7a4f0;
+}
+
+.book-card img {
+    height: 200px;
+    width: 150px;
+    margin: 5px;
+    margin-bottom: 2px;
+}
+
+.book-title {
+    margin: 2px 5px;
+    word-wrap: normal;
+}
+
+.book-author {
+    margin: 2px 5px;
+    font-weight: bold;
+    font-size: x-small;
+    word-wrap: normal;
+}
+
+.book-description {
+    margin: 2px;
+    margin-bottom: 5px;
+    font-size: small;
+    overflow-y: scroll;
+
+}
