+++ frontend/src/components/BookList/BookList.css
@@ -1,10 +1,5 @@
 .books-page {
     display: flex;
     flex-direction: column;
-}
-
-.books-list {
-    display: flex;
-    flex-direction: column;
-    overflow-y: scroll;
+    width: 100%;
 }
