+++ frontend/src/components/BookList/index.js
@@ -1,24 +1,16 @@
 import { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import * as bookActions from "../../store/book";
-import { useHistory } from "react-router-dom/cjs/react-router-dom.min";
 import './BookList.css';
+import BookCard from "../BookCards";
 
 const BookList = () => {
     const dispatch = useDispatch();
-    const history = useHistory();
     const list = useSelector(bookActions.selBooks);
 
     const [books, setBooks] = useState([]);
     const [author, setAuthor] = useState('');
     const [title, setTitle] = useState('');
-
-
-    const handleClick = (id) => (e) => {
-        e.preventDefault();
-
-        history.push(`/books/${id}`);
-    };
 
     const handleChange = (e) => {
         e.preventDefault();
@@ -62,16 +54,7 @@ const BookList = () => {
             <input type='text' className='title' placeholder="Search by title" onChange={handleChange} value={title} />
             <input type='text' className='author' placeholder="Search by author" onChange={handleChange} value={author} />
 
-            <div className='books-list' >
-            {books.length > 0 ? books.map(book => (
-                <div className="book-card" onClick={handleClick(book.id)} key={book.id}>
-                    <img src={book.thumbnail} alt={book.title} />
-                    <h4 className="book-title">{book.title}</h4>
-                    <p className="book-author">{book.authors}</p>
-                    <p className="book-description">{book.synopsis}</p>
-                </div>
-            )) : <h1>Hmm, it seems we ran out of books...</h1>}
-            </div>
+            <BookCard books={books} />
         </div>
     )
 }
