+++ frontend/src/helpers/index.js
@@ -1,3 +1,5 @@
+import defaultImage from '../assets/no-book-found.png';
+
 export const isValidUrl = urlString => {
     try {
         return Boolean(new URL(urlString));
@@ -6,3 +8,9 @@ export const isValidUrl = urlString => {
         return false;
     }
 };
+
+
+export const setDefaultImg = (e) => {
+    e.target.onerror = null;
+    e.target.src = defaultImage;
+}
