+++ frontend/src/index.css
@@ -2,6 +2,7 @@
 .app {
     display: flex;
     max-height: 100vh;
+    max-width: 100vw;
 }
 
 .navbar {
@@ -17,10 +18,7 @@
     position: relative;
 }
 
-.content {
-    flex: 1 1 auto;
-    padding: 10px;
-}
+
 
 html,
 body,
@@ -34,3 +32,4 @@ body,
 i {
     color: #8b0280;
 }
+
